Update Query.py
Changes to a number of the Query functions to work around an issue with Catalyst optimization.  Catalyst will get confused if you pass in DataFrames that are derived from a common parent DataFrame.  The changes will create an intermediate Dataframe with the columns renamed to get around the issue.  Believe it is related to this issue:  https://issues.apache.org/jira/browse/SPARK-10925